+++ controllers/usuariosController.js
@@ -1,25 +1,6 @@
 const usuariosRepository = require('../repositories/usuariosRepository');
 const ApiError = require('../utils/errorHandler');
 
-const getUsuarios = async (req, res, next) => {
-  try {
-    const usuarios = await usuariosRepository.findAll();
-    res.status(200).json(usuarios);
-  } catch (error) {
-    next(new ApiError('Erro ao obter os usuarios', 500, error.message));
-  }
-};
-
-const getUsuarioById = async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const usuario = await usuariosRepository.findById(id);
-    res.status(200).json(usuario);
-  } catch (error) {
-    next(new ApiError('Erro ao obter o usuario', 500, error.message));
-  }
-};
-
 const getMe = async (req, res, next) => {
   try {
     const userId = req.user.id;
@@ -39,9 +20,34 @@ const getMe = async (req, res, next) => {
   }
 };
 
+const getUsuarios = async (req, res, next) => {
+  try {
+    const usuarios = await usuariosRepository.findAll();
+    res.status(200).json(usuarios);
+  } catch (error) {
+    next(new ApiError('Erro ao obter os usuarios', 500, error.message));
+  }
+};
+
+const getUsuarioById = async (req, res, next) => {
+  try {
+    const id = Number(req.params.id);
+    if (isNaN(id)) {
+      return next(new ApiError('ID inválido', 400));
+    }
+    const usuario = await usuariosRepository.findById(id);
+    res.status(200).json(usuario);
+  } catch (error) {
+    next(new ApiError('Erro ao obter o usuario', 500, error.message));
+  }
+};
+
 const deleteUsuario = async (req, res, next) => {
   try {
-    const { id } = req.params;
+    const id = Number(req.params.id);
+    if (isNaN(id)) {
+      return next(new ApiError('ID inválido', 400));
+    }
     const deletedUsuario = await usuariosRepository.remove(id);
     res.status(200).json({
       message: 'Usuario excluido com sucesso',
